Refactor thumbnail generation
